+++ scanpipe/views.py
@@ -498,11 +498,9 @@ class PaginatedFilterView(FilterView):
         context["url_params_without_page"] = query_dict.urlencode()
 
         context["searchable_fields"] = sorted(
-            [
-                field.name
-                for field in self.model._meta.get_fields()
-                if not field.is_relation
-            ]
+            field.name
+            for field in self.model._meta.get_fields()
+            if not field.is_relation
         )
 
         return context